feat(db/hooks): port over special rarity implementation to showcase sample usage of hooks
